requirements.txt
attrs==22.2.0
exceptiongroup==1.1.0
iniconfig==2.0.0
iota-client==1.0.0rc2
iota-wallet @ file:///home/antonio/dev/crew3_shimmer_token_sender/wallet.rs/bindings/python
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
